Added most of the code to browse objects Added some tests that fail.
Now, after a few hours of debugging i reached the conclusion that the ids provided by the devTools are not for the objects but for proxies that it creates on each request, this means that i have no way of mantaining identity between different requests of data, at least for now.

All my tests will have to be by value.